requirements.txt
flask
openai-whisper
cloudinary
google-generativeai
edge-tts
git+https://github.com/openai/whisper.git
moviepy===1.0.3
requests
python-dotenv
tempfile
requests
gc
os
re
